add benches/setup
